refactor file system storage

+++ src/server/utils/uploadCrudFunctions.js
@@ -1,16 +1,17 @@
 var fs = require("fs");
-var config = require("../../../config.js");
 
 exports.get = (req,res) => {
 	crudFunctions(req,res);
 }
 
 exports.post = (req,res) => {
-	const { folderName, folder , images} = req.body;
+	const { folderName,folder,images } = req.body;
+	// if post request has images/folder data, add that data to the system.
 	if (folder||images){
 		crudFunctions(req,res,addFunc);
 	}else{
-		crudFunctions(req,res,getPhoto);
+	//
+		crudFunctions(req,res,getBoth);
 	}
 }
 
@@ -30,26 +31,22 @@ const crudFunctions = (req,res,action) => {
 var filePath = './src/client/public/images/uploads/';
 var imagePath = '../images/uploads/';
 
-//get all folders and photos from file storage system.
-const getPhoto = (req,res) => {
-	const {folderName} = req.body;
+//get all folders and specific photos from file storage system.
+const getBoth = (req,res) => {
+	const { folderName } = req.body;
 	let folders = fs.readdirSync(filePath);
-	let photos = fs.readdirSync(filePath+"/"+folderName);
-	
-	
+	let photos = fs.readdirSync(filePath+"/"+folderName);	
 	let action1 = folders.map(folder => ({name:folder,path:imagePath+folder}));
 	let action2 = photos.map(photo => ({name:photo,path:imagePath+'/'+folderName+'/'+photo}));
-	
 	let promiseArray = [Promise.resolve(action1),Promise.resolve(action2)];	
 	Promise.all(promiseArray).then(data => {
 		res.json({images:data[1],folders:data[0],folderName:folderName});
 	}).catch(error => {
 		console.log("error from getFiles" +error);
-	})
-	
+	})	
 }
 
-//get all folders and photos from file storage system.
+//get all folders 
 const getFolder = (req,res) => {
 	let files = fs.readdirSync(filePath);
 	let action = files.map(file => ({name:file,path:imagePath+file}));
@@ -70,31 +67,18 @@ const deleteFunc = (req,res) => {
 			if (error){
 				console.log("error from deletePhoto " + error);
 			}
-			getPhoto(req,res)
+			getBoth(req,res)
 		})
-	}else{
+	}else if(type === 'folder'){
 		fs.rmdir(filePath+data,(error) => {
 			if(error){
 				console.log("error from deleteFolder " + error);
 			}
-			getFunc(req,res)
+			getFolder(req,res)
 		})
 	}
 }
 
-
-//get all folders and photos from file storage system.
-const getFunc = (req,res) => {
-	let files = fs.readdirSync(filePath);
-	let action = files.map(file => ({name:file,path:imagePath+file}));
-	let promise = new Promise((resolve,reject) => {
-		resolve(action);
-	}).then(data => {
-		res.json({folders:data});
-	}).catch(error => {
-		console.error("error from getFiles " + error);
-	})
-}
 //add item to file storage system.
 const addFunc = (req,res) => {
 	const { images,folder,folderName } = req.body;
@@ -106,7 +90,7 @@ const addFunc = (req,res) => {
 				if (error){
 					console.log("error from add Photos "+ error);
 				}
-				getPhoto(req,res);
+				getBoth(req,res);
 			});
 		})
 	}else if(folder){
@@ -114,7 +98,7 @@ const addFunc = (req,res) => {
 			if(error){
 				console.log("error from add folder " + error);
 			}
-			getFunc(req,res);
+			getFolder(req,res);
 		});	
 	}	
 }
